Return coinstake as protobuf

+++ cypcore/Models/SendPaymentProto.cs
@@ -9,13 +9,13 @@ namespace CYPCore.Models
     public class SendPaymentProto
     {
         [ProtoMember(1)]
-        public double Amount { get; set; }
+        public ulong Amount { get; set; }
         [ProtoMember(2)]
         public string Address { get; set; }
         [ProtoMember(3)]
         public CredentialsProto Credentials { get; set; }
         [ProtoMember(4)]
-        public double Fee { get; set; }
+        public ulong Fee { get; set; }
         [ProtoMember(5)]
         public string Memo { get; set; }
         [ProtoMember(6)]
